Move next show to nav
