+++ src/components/hero/hero.js
@@ -1,54 +1,9 @@
 import * as React from 'react'
 import { hero } from './hero.module.css';
-import { useStaticQuery, Link, graphql } from 'gatsby';
 
 const Hero = () => {
-  const nextShowQuery = useStaticQuery(graphql`
-    query {
-      allContentfulShow(sort: {date: DESC}, limit: 1) {
-        edges {
-          node {
-            id
-            title
-            slug
-            location
-            ticketsUrl
-            description {
-              raw
-            }
-            date
-            image {
-              title
-              file {
-                fileName
-                url
-              }
-            }
-          }
-        }
-      }
-    }
-  `);
-
-  const now = new Date();
-  const shows = nextShowQuery.allContentfulShow.edges.map(edge => edge.node);
-  const nextShow = shows.find(show => new Date(show.date).setUTCHours(23,59,59,999) > now);
-
-  const nextShowTemplate = () => {
-    if (nextShow) {
-      return (
-        <Link to={`/shows/${nextShow.slug}`}>
-          <p>{nextShow.title}</p>
-        </Link>
-      );
-    } else {
-      return null;
-    }
-  }
-
   return (
     <div className={hero}>
-      { nextShowTemplate() }
       <svg xmlns="http://www.w3.org/2000/svg" version="1.1" viewBox="0 0 557.87 297.26">
         <g id="background">
           <path d="M418.72,104.42c2.52-4.29,.07-9.96-3.47-13.02-1.03-.89-2.35-1.65-3.69-1.38-1.28,.26-2.17,1.38-2.92,2.45-1.86,2.69-3.48,5.55-5.55,8.09-9.83,12.07-31.29,6.65-36.79-7.21-1.09-2.74-1.74-5.65-2.12-8.65,1.34-9.82,2.32-19.69,3.1-29.57,5.97-1.34,11.5-3.7,17.62-4.81s12.82-2.04,18.05-5.97c.7-.53,1.39-1.13,1.73-1.95,.64-1.57-.3-3.43-1.7-4.38-1.4-.95-3.15-1.22-4.83-1.46-5.86-.84-11.94-1.72-17.84-.72-1.07,.18-11.9,2.59-11.92,2.88,.6-10.25,1.05-20.51,1.49-30.76,.44-10.18-14.87-10.82-15.74-.75l-3.06,35.41-16.93,3.81c-9.89,2.22-6.45,17.6,3.45,15.37l12.06-2.71c-.04,.47-.08,.93-.12,1.4-.64,7.39-1.39,15.17-1,22.76-.21,1.59-.41,3.18-.64,4.77-1.68,11.47-3.02,25.13-9.66,34.99-4.95,7.36-16.38,13.76-23.53,5.45-8.27-9.6-6.23-52.17-29.37-39.26-4.88,2.72-9.18,6.46-12.88,10.85,.02-1.07,.05-2.13,.03-3.21-.24-10.04-15.55-10.98-15.74-.75-.03,1.77-.07,3.55-.1,5.32-.88,16.26-7.52,31.77-19.61,43.04-2.98,2.78-7.96,7.7-12.36,5.19-6.09-3.47-4.42-10.62-3.36-15.95,1.38-6.92,2.17-13.95,2.31-21.01,.02-.88,0-1.84-.53-2.55-.69-.92-1.99-1.08-3.14-1.17-3.88-.29-8.71-.06-10.38,3.45-.54,1.13-.62,2.4-.69,3.65-.69,11.95-1.3,24.48-7.43,34.76-1.45,2.44-3.33,4.81-5.98,5.83-1.55,.6-3.25,.69-4.91,.68-4.11-.01-8.22-.56-12.19-1.63-1.39-.37-2.83-.85-3.8-1.91-1.48-1.6-1.47-4.09-.94-6.2,1.17-4.64,.67-9.18,1.46-13.96,.9-5.44,2.54-10.81,2.35-16.38-.02-.74-.09-1.5-.45-2.14-.33-.59-.89-1.01-1.47-1.35-4.24-2.49-8.41-.52-10.98,3.3-1.12,1.66-1.8,3.56-2.47,5.44-8.11-2.15-17.43-.28-23.97,4.95-8.43,6.74-16.02,14.69-21.76,23.87-6.55,10.46-14.7,19.29-27.17,22.63-3.94,1.06-7.96,1.65-12,1.94,16.9-25.56,24.69-57.99,29.49-88.05,1.98-12.4,4.22-25.7-1.41-37.53-4.52-9.49-14.9-16.78-25.63-15.83-12.21,1.09-18.49,11.73-22.92,21.88-4.72,10.79-8.42,22.04-11.17,33.49-5.45,22.73-7.27,46.32-5.44,69.62,.5,6.41,7.01,28.08,2.7,32.26-1.16,1.13-8.74,.9-10.56,1.34-5.47,1.32-10.93,3.51-15.78,6.34-7.29,4.26-13.57,10.76-16.08,18.83-.64,2.07-.63,4.99,1.42,5.68,1.43,.48,2.91-.5,4.06-1.48,7.28-6.16,13.9-14.48,23.37-15.67,4.93-.62,16.82-2.74,19.74-.13,4.38,3.93,6.89,19.69,9.47,25.49,8.72,19.6,20.92,41.49,40.78,51.62,17.46,8.9,39.79,5.91,52.04-10.33,17.6-23.32-7.41-47.4-24.62-61.71-10.58-8.79-22.11-16.34-34.33-22.5,7.69,0,14.24-6.35,19.45-11,1.2-1.07,2.58-2.21,4.19-2.14,2.23,.09,3.61,2.34,5.2,3.9,4.04,3.97,10.46,3.9,16.07,3.2,9.48-1.19,20.13-4.46,23.97-13.21,5.42,5.13,13.47,6.69,20.88,5.84,7.25-.84,15.59-5.14,22.06-8.28,3.75-1.82,7.07-4.76,10.97-5.42,7.75-1.31,17.15,2.34,24.28-.9s13.9-12.63,17.45-19.61c-2.41,4.75,14.49,22.07,17.14,8.33,.44-2.3,1.3-11.73,1.07-14.06-.79-8,.91-16.23,4.82-23.26,1.18-2.12,2.64-4.21,4.77-5.39s5.07-1.16,6.75,.59c1.02,1.06,1.42,2.56,1.76,3.99,2.78,11.85,3.16,22.43,13.54,30.67,8.05,6.39,18.55,7.01,27.9,3.24,8.25-3.33,13.5-10.33,16.76-18.51,.96-2.4,1.67-4.91,2.98-7.14s3.35-4.2,5.89-4.65c2.14-.38,4.3,.36,6.38,.98,14.21,4.25,30.17,4.11,42.98-3.65,2.97-1.8,7.07-4.75,9.16-8.32ZM59.85,118.03c1.64-15.42,5.56-30.57,11.58-44.86,2.54-6.03,7.18-14.68,14.1-7.68,5.85,5.91,3.49,16.41,2.37,23.69-3.89,25.2-10.2,53.38-26.64,73.24-.07,0-.14-.01-.21-.02-2.36-14.64-2.76-29.61-1.2-44.37Zm43.04,98.82c6.51,4.34,12.61,9.22,18.24,14.66,5.16,4.99,12.95,11.7,13.08,19.46,.2,11.93-16.12,16.17-25.13,12.87-13.73-5.02-21.98-19.71-28.08-32.07-4.79-9.7-8.52-19.87-11.24-30.3,.17-.16,.34-.32,.51-.48,11.52,3.81,22.47,9.07,32.64,15.86Zm54.77-77.44c-.32,1.09-.65,2.07-.76,2.58-.5,2.41-.9,4.79-.91,7.23-2.95,8.53-9.69,15.66-19.23,16.53-8.76,.8-6.26-6.17-2.84-11.7,1.02-1.51,2.07-3.01,3.14-4.46,.2-.25,.39-.48,.56-.68,5.39-6.39,12.08-11.22,19.52-14.87,.35,1.75,.53,3.55,.53,5.37Z"/>
